Start using Cocoapods
Cocoapods makes managing frameworks a lot easier. It acts as a package
manager and updates any frameworks that aren't up-to-date.

Not all frameworks are available on Cocoapods, so those need to be
installed manually. Overuse of Cocoapods can lead to needing a lot of
dependencies, which is bad practice. Tread lightly.