Fix crashing when itemURL is nil
If the user had a login item that pointed to a nonexistent file, itemURL
would return nil. By calling takeRetainedValue on nil, the program would
crash for unwrapping a nil value.

A check is done on whether itemURL is nil or not. If it is, then it
won't run takeRetainedValue and will just finish the iteration.
Otherwise, the file to which the URL refers to exists and can be checked
against the bundle's URL.